requirements.txt
# Aura Web Studio — Python dependencies
# Install on PC: pip install -r requirements.txt

# ADK + Gemini
google-adk>=1.0.0
google-generativeai>=0.8.0

# API backend
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0

# Utilities
python-dotenv>=1.0.0
